+++ src/Akeneo/Category/back/Domain/Query/GetCategoryTreeTemplates.php
@@ -8,10 +8,13 @@ use Akeneo\Category\Domain\ValueObject\CategoryId;
 use Akeneo\Category\Domain\ValueObject\Template\TemplateUuid;
 
 /**
- * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
+ * @copyright 2023 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  */
-interface DeleteCategoryTreeTemplateByCategoryIdAndTemplateUuid
+interface GetCategoryTreeTemplates
 {
-    public function __invoke(CategoryId $categoryTreeId, TemplateUuid $templateUuid): void;
+    /**
+     * @return TemplateUuid[]
+     */
+    public function __invoke(CategoryId $categoryTreeId): array;
 }